inf-spring-blog-hodolman/src/main/java/com/ys/springboot_blog_hodol/service/AuthService.java
package com.ys.springboot_blog_hodol.service;

import java.util.Optional;

import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import com.ys.springboot_blog_hodol.crypto.PasswordEncoder;
import com.ys.springboot_blog_hodol.domain.User;
import com.ys.springboot_blog_hodol.exception.AlreadyExistsEmailException;
import com.ys.springboot_blog_hodol.exception.InvalidSigninInformation;
import com.ys.springboot_blog_hodol.repository.UserRepository;
import com.ys.springboot_blog_hodol.request.Login;
import com.ys.springboot_blog_hodol.request.Signup;

import lombok.RequiredArgsConstructor;

@Service
@RequiredArgsConstructor
public class AuthService {

    private final UserRepository userRepository;

    @Transactional
    public Long signin(Login login) {
        User user = userRepository.findByEmail(login.getEmail())
                .orElseThrow(InvalidSigninInformation::new);

        PasswordEncoder encoder = new PasswordEncoder();
        var matches = encoder.matches(login.getPassword(), user.getPassword());
        if (!matches) {
            throw new InvalidSigninInformation();
        }

        return user.getId();
    }

    public void signup(Signup signup) {
        Optional<User> userOptional = userRepository.findByEmail(signup.getEmail());
        if (userOptional.isPresent()) {
            throw new AlreadyExistsEmailException();
        }

        PasswordEncoder encoder = new PasswordEncoder();
        String encryptedPassword = encoder.encrpyt(signup.getPassword());

        var user = User.builder()
                .email(signup.getEmail())
                .password(encryptedPassword)
                .name(signup.getName())
                .build();
        userRepository.save(user);
    }
}
